==> code/01_load_postgres_db.ipynb <==
"""

load CSV files downloadedfrom SEC database queries in Google Public Cloud
into postgres db created in local drive

"""

from sqlalchemy import create_engine
import pandas as pd


# local db in postgres is named 'bankrupcy'

engine = create_engine('postgresql://amybutler:localhost@localhost:5432/bankruptcy')

## add table with list of industry codes

sic_codes = pd.read_csv('../data/SIC_codes.csv', dtype={'SIC Code':'str'})

sic_codes.to_sql('sic_codes', engine, index=False)

query = 'SELECT * FROM sic_codes;'
df = pd.read_sql(query, engine)

df.head()

# add table with list of industry divisions


sic_divisions = pd.read_csv('../data/SIC_divisions.csv', dtype={'div_code':'str'})

sic_divisions.to_sql('sic_divisions', engine, index=False)

query = 'SELECT * FROM sic_divisions;'
df = pd.read_sql(query, engine)

df.head()

# Annual Reports in 10-K  and 10-K/A (ammendments to 10-K)
# CSV's downloaded from querying the Google Cloud Platform SEC Public Dataset 
# SEC is U.S. Securities and Exchange Commission

file1 = '../data/2014_SEC_10K_raw.csv'
file2 = '../data/2015_SEC_10K_raw.csv'
file3 = '../data/2016_SEC_10K_raw.csv'
file4 = '../data/2017_SEC_10K_raw.csv'
file5 = '../data/2018_SEC_10K_raw.csv'
file6 = '../data/2019_SEC_10K_raw.csv'

file1_df = pd.read_csv(file1, dtype={'sic':  'str', 'ein':'str', 'central_index_key':'str'}, 
                        parse_dates=['period_end_date', 'date_filed'] )

file2_df = pd.read_csv(file2, dtype={'sic':  'str', 'ein':'str', 'central_index_key':'str'}, 
                        parse_dates=['period_end_date', 'date_filed'] )

file3_df = pd.read_csv(file3, dtype={'sic':  'str', 'ein':'str', 'central_index_key':'str'}, 
                        parse_dates=['period_end_date', 'date_filed'] )

file4_df = pd.read_csv(file4, dtype={'sic':  'str', 'ein':'str', 'central_index_key':'str'}, 
                        parse_dates=['period_end_date', 'date_filed'] )


file5_df = pd.read_csv(file5, dtype={'sic':  'str', 'ein':'str', 'central_index_key':'str'}, 
                        parse_dates=['period_end_date', 'date_filed'] )

file6_df = pd.read_csv(file6, dtype={'sic':  'str', 'ein':'str', 'central_index_key':'str'}, 
                        parse_dates=['period_end_date', 'date_filed'] )

print('file1: 2014', file1_df.shape)
print('file2: 2015', file2_df.shape)
print('file3: 2016', file3_df.shape)
print('file4: 2017', file4_df.shape)
print('file5: 2018', file5_df.shape)
print('file6: 2019', file6_df.shape)


# the difference in number of rows later determined to be caused by many companies
# keep re-sending their data to SEC when they send the most recent report

# also 2019 Q4 is not available in Google Cloud. Is available in a download file from SEC website. 
# will prepare those files separately 

file1_df.to_sql('sec_raw_2014', engine, index=False)

file2_df.to_sql('sec_raw_2015', engine, index=False)
file3_df.to_sql('sec_raw_2016', engine, index=False)
file4_df.to_sql('sec_raw_2017', engine, index=False)
file5_df.to_sql('sec_raw_2018', engine, index=False)
file6_df.to_sql('sec_raw_2019', engine, index=False)


"""
In terminal, in psql connected to local postgres db, 
the 6 raw SEC data tables in postgres db were combined by:


CREATE TABLE sec_raw_all as 
SELECT * FROM sec_raw_2014
UNION
SELECT * FROM sec_raw_2015
UNION
SELECT * FROM sec_raw_2016
UNION
SELECT * FROM sec_raw_2017
UNION
SELECT * FROM sec_raw_2018
UNION
SELECT * FROM sec_raw_2019;


total rows = 1859118


"""

# load tables with 10-Q and 10-Q/A quarterly reports

# CSV's downloaded from querying the Google Cloud Platform SEC Public Dataset 
# SEC is U.S. Securities and Exchange Commission

file7  = '../data/2014_SEC_10Q_raw.csv'
file8  = '../data/2015_SEC_10Q_raw.csv'
file9  = '../data/2016_SEC_10Q_raw.csv'
file10 = '../data/2017_SEC_10Q_raw.csv'
file11 = '../data/2018_SEC_10Q_raw.csv'
file12 = '../data/2019_SEC_10Q_raw.csv'

file7_df = pd.read_csv(file7, dtype={'sic':  'str', 'ein':'str', 'central_index_key':'str'}, 
                        parse_dates=['period_end_date', 'date_filed'] )

file8_df = pd.read_csv(file8, dtype={'sic':  'str', 'ein':'str', 'central_index_key':'str'}, 
                        parse_dates=['period_end_date', 'date_filed'] )

file9_df = pd.read_csv(file9, dtype={'sic':  'str', 'ein':'str', 'central_index_key':'str'}, 
                        parse_dates=['period_end_date', 'date_filed'] )

file10_df = pd.read_csv(file10, dtype={'sic':  'str', 'ein':'str', 'central_index_key':'str'}, 
                        parse_dates=['period_end_date', 'date_filed'] )

file11_df = pd.read_csv(file11, dtype={'sic':  'str', 'ein':'str', 'central_index_key':'str'}, 
                        parse_dates=['period_end_date', 'date_filed'] )

file12_df = pd.read_csv(file12, dtype={'sic':  'str', 'ein':'str', 'central_index_key':'str'}, 
                        parse_dates=['period_end_date', 'date_filed'] )

print(file7_df.shape)
print(file8_df.shape)
print(file9_df.shape)
print(file10_df.shape)
print(file11_df.shape)
print(file12_df.shape)

file7_df.to_sql('sec_10q_raw_2014', engine, index=False)


file8_df.to_sql('sec_10q_raw_2015', engine, index=False)


file9_df.to_sql('sec_10q_raw_2016', engine, index=False)


file10_df.to_sql('sec_10q_raw_2017', engine, index=False)


file11_df.to_sql('sec_10q_raw_2018', engine, index=False)


file12_df.to_sql('sec_10q_raw_2019', engine, index=False)
