==> code/02_EDA_SEC_data.ipynb <==
"""

EDA of TRAINING and VALIDATION DATA
SEC filings 2014-2018

Downloaded from Google Cloud and saved in local Postgres database

- amount of missing data points in SEC data
- records reported in non-USD


"""

import psycopg2 as pg
import pandas.io.sql as pd_sql
import pandas as pd
import numpy as np
import datetime


# Postgres info to connect

connection_args = {
    'host': 'localhost',  
    'dbname': 'bankruptcy',  
    'port': 5432          
}

connection = pg.connect(**connection_args)
connection.autocommit = True   

# few fields pulled

query = """
SELECT distinct measure_tag
FROM sec_all_raw
WHERE  period_end_date < '2019-01-01' 
    and units = 'USD'

;
"""

sec_data = pd_sql.read_sql(query, connection)
sec_data.head()

len(sec_data)

sec_data.tail()

all_columns = sec_data.transpose()



all_columns

#########

# Which companies are submitting filings with non-USD values

query = """
SELECT distinct company_name, units
FROM sec_all_raw
WHERE  period_end_date < '2019-01-01' 
    and units != 'USD'

;
"""

sec_data = pd_sql.read_sql(query, connection)
sec_data.head()


len(sec_data)

sec_data.tail()

## After researching, found non-U.S. public companies can and do file the 10-K
## Though supposed to use a different SEC form 20-F
## there doesn't seem to be a requirement that they not use the 10-K

## will remove the companies with non-USD units from data set

#################

#Revenue can be in a few different fields, may need to pull more data from Google Cloud

query = """
SELECT company_name, period_end_date, measure_tag, value
FROM sec_all_raw
WHERE  period_end_date < '2019-01-01' 
    and units = 'USD'
    and measure_tag in ('ProfitLoss', 'Revenues', 'SalesRevenueNet', 'Assets', 'GrossProfit')

;
"""

sec_data = pd_sql.read_sql(query, connection)
sec_data.head()

pivoted = sec_data.pivot_table(index=['company_name', 'period_end_date'], columns= 'measure_tag',
                                values='value').reset_index()

over_100 = pivoted[pivoted['Assets']>= 100000000]

len(over_100)

over_100.info()

over_100[over_100['company_name'] == 'ALPHABET INC.']

over_100[over_100['company_name'] == 'AMAZON COM INC']

over_100[(over_100['Revenues'].isna() ) & (over_100['SalesRevenueNet'].isna() ) ]

###################

# Examine companies with many blank fields

query = """
SELECT *
FROM sec_wide_all
WHERE  period_end_date < '2019-01-01' 
;
"""

sec_data = pd_sql.read_sql(query, connection)
sec_data.head()

sec_data.columns

sec_data.info()

# look at companies who don't have "Assets" filled in, determine if company using another field
sec_data[sec_data['Assets'].isna()].company_name.unique()

examine_one = sec_data[sec_data['company_name']=='ZYNGA INC'].sort_values(by='period_end_date')
examine_one.to_excel('examine_one_co.xlsx', index=False)

# ZYNGA keeps re-submitting the annual report for a previous end_date.  Not different data, 
# just less data than originally submitted

# will examine another company
examine_two = sec_data[sec_data['company_name']=='ZZLL INFORMATION TECHNOLOGY, INC'].sort_values(by='period_end_date')
examine_two.to_excel('examine_two_co.xlsx', index=False)

## ZZLL submitted updated information

# 12 RETECH CORP

examine_3 = sec_data[sec_data['company_name']=='12 RETECH CORP'].sort_values(by='period_end_date')
examine_3.to_excel('examine_3_co.xlsx', index=False)

# resubmitting prior period_end the following year but not submitting new informaton.
