==> code/03_BRD_web_scrape.ipynb <==
""" 
code to scrape list of large public company bankruptcies from web page: 
"https://lopucki.law.ucla.edu/spreadsheet.htm"

the UCLA School of Law Bankruptcy Research Databasee

where "large"  includes companies with reported assets over $100MM (in 1980 dollars)

will re-visit amending this list to include all public company bankruptcies 2014-2020
"""

from bs4 import BeautifulSoup
import requests

import numpy as np
import pandas as pd


url = 'https://lopucki.law.ucla.edu/spreadsheet.htm' 

response = requests.get(url)

response.status_code  #200 = success!

response.text[:1000]  #First 1000 characters of the HTML

page = response.text

soup = BeautifulSoup(page)

table = soup.find(id='xyscrolling').find_all('td')

#table[:250]

contents_table = [x.text.strip() for x in table]

len(contents_table)

data_array = np.array(contents_table).reshape(1208, 25)

data_array[1]

brd_list = pd.DataFrame(data_array)
brd_list.columns = ['debtor_name', 'industry_division', 'industry_group',
                    'cause', 'date_filed', 'disposition', '_363_sale',
                   'duration', 'inc_state', 'district_filed', 'city_filed',
                   'forum_shopping', 'filing_party', 'plan_type', 'employees',
                   'business_survival', 'refiled', 'hq_city', 'date_disposed',
                   'asset_size', 'employee_size', 'annual_sales', 'assets_$MM',
                   'chapter_filing', 'debtor_attorney']

brd_list.head()

brd_list.shape

# force the couple date columns to be in date format for easy querying later
brd_list['date_filed']= pd.to_datetime(brd_list['date_filed'])
brd_list['date_disposed']= pd.to_datetime(brd_list['date_disposed'])

brd_list.head()

## add table to postgres db
from sqlalchemy import create_engine

engine = create_engine('postgresql://amybutler:localhost@localhost:5432/bankruptcy')

brd_list.to_sql('ucla_brd_list', engine, index=False)

# check table saved and can query
query = 'SELECT * FROM ucla_brd_list;'
df = pd.read_sql(query, engine)

df.head()
