==> code/04_BRD_list_prep.ipynb <==
"""
filter BRD list to date_filed between 1/1/2014 - 12/31/2020

Found many EIN missing from SEC data, so prepared list of bankruptcies
to have the same name as in the company_name field of SEC data

Add labeling of target field (bankruptcy w/in 1 year of annual report = 1 else 0)


"""

import psycopg2 as pg
import pandas as pd
import pandas.io.sql as pd_sql
import numpy as np
import datetime

from pandasql import sqldf

# PandaSQL needs to be able to reference the global variables already defined (namely, `df`)
pysqldf = lambda q: sqldf(q, globals())

# Postgres info to connect

connection_args = {
    'host': 'localhost',  
    'dbname': 'bankruptcy',  
    'port': 5432          
}

connection = pg.connect(**connection_args)
connection.autocommit = True   

query = """
SELECT debtor_name, date_filed, chapter_filing 
FROM ucla_brd_list 
WHERE (date_filed >= '2015-01-01') and (date_filed  <= '2020-12-31')
;"""

brd_query =  pd_sql.read_sql(query, connection)

brd_query.head()

brd_2014.shape

query_sec = """
SELECT distinct company_name, ein
FROM sec_all_usd
"""
companies = pd_sql.read_sql(query_sec, connection)

companies.head()

query_one = """
SELECT *
FROM sec_all_usd
WHERE company_name = 'USEC INC';
"""

one_query = pd_sql.read_sql(query_one, connection)

one_query.head()

## first try join after uppercase debtor_name in brd list

join_try1 = """
SELECT distinct a.debtor_name, a.date_filed, a.chapter_filing, b.ein
FROM ucla_brd_list as A
    LEFT JOIN sec_all_usd as B
ON UPPER(A.debtor_name) = UPPER(B.company_name)
WHERE a.date_filed >= '2014-01-01'
;"""

t1_join = pd_sql.read_sql(join_try1, connection)
t1_join.head()

t1_join.shape

t1_join[t1_join['ein'].notna()].count()

t1_join[t1_join['ein'].isna()].count()

## export t1_join and look up eins

t1_join.to_excel('debtor_list_ein_review.xlsx', index=False)

# Query db table to find EINS 
# This query was run mulitple times to fill in missing EINS
# where not name match exactly between the two data sources

query_co = """
SELECT distinct ein, company_name
FROM sec_all_usd
WHERE lower(company_name) LIKE '%melinta%'
;
"""

find_ein = pd_sql.read_sql(query_co, connection)

find_ein
