==> code/05_SEC_data_prep.ipynb <==
"""
Prepare SEC data to be analyzed further and for classification model features.
As-is all the financial statement labels in the column "measure_tag"

The SEC data is tall and needs pivoted so model feature in own column

Apply additional logics to fill in where multiple possible columns
for same measure. There is not any standardization as to what measures 
each company has to submit or how the measures be labeled. 

Apply logics to fill forward where multiple submissions for any period_end_date
and the company did not re-submit previously provided data

Keep only one row for any company and period_end_date of financial measures


Add the industry code descriptions 
Add the industry division descriptions

"""

# based on EDA, will filter to only where USD values, 
# some non-US companies sent SEC the US form 10-K

# this code was run to create separate table in local postgres db

"""  
CREATE TABLE sec_usd_all as 
SELECT * 
FROM sec_ALL_raw 
WHERE units='USD';


rows:1856236
"""

import psycopg2 as pg
import pandas as pd
import pandas.io.sql as pd_sql
import numpy as np
import datetime as dt

from pandasql import sqldf
# PandaSQL needs to be able to reference the global variables already defined
pysqldf = lambda q: sqldf(q, globals())

# Postgres info to connect

connection_args = {
    'host': 'localhost',  
    'dbname': 'bankruptcy',  
    'port': 5432          
}

connection = pg.connect(**connection_args)
connection.autocommit = True   
                              

query = "SELECT * FROM sec_usd_all LIMIT 5;"

pd_sql.read_sql(query, connection)

query_all = "SELECT * FROM sec_usd_all;"
sec_df = pd_sql.read_sql(query_all, connection)

sec_df.shape

sec_df.info()

sec_df.columns

# from EDA of raw data, filter to where number of quarters is 0 or 4
# the 10-K are annual reports

sec_df2 = sec_df[sec_df['number_of_quarters'].isin([0,4])]
sec_df2.shape

#pivot the tall df to be wide on the measures

wide_sec = sec_df2.pivot_table(index=['company_name','period_end_date', 'number_of_quarters',
                                              'submission_number','central_index_key',
                                             'ein', 'sic', 'fiscal_year_end', 'fiscal_year', 'form',
                                             'date_filed'],
                        columns='measure_tag', values='value').reset_index()
wide_sec.head()

wide_sec['number_of_quarters'].unique()


wide_sec.shape

# drop any finanical report data before 2014

f1_wide_sec = wide_sec[wide_sec['period_end_date'] > dt.datetime(2014, 1, 1) ]
f1_wide_sec.shape

f1_wide_sec.columns

# sort df
f1_wide_sec.sort_values(by=['company_name', 'period_end_date', 'date_filed'], inplace=True)
f1_wide_sec.head()


# From EDA of SEC data, some companies keep re-subitting previous reports,
# sometimes with only one new value, need to get the previously submitted data

# fill forward within groupby

# leave the NaN's to apply additional logics

cols = ['Assets',
       'AssetsCurrent', 'CashAndCashEquivalentsAtCarryingValue',
       'CashAndCashEquivalentsPeriodIncreaseDecrease', 'CommonStockValue',
       'CostsAndExpenses', 'CurrentAsset', 'DeferredIncomeTaxExpenseBenefit',
       'DeferredIncomeTaxLiabilities', 'DeferredIncomeTaxLiabilitiesNet',
       'DeferredIncomeTaxesAndTaxCredits', 'Depreciation',
       'EarningsPerShareBasic', 'EarningsPerShareDiluted', 'Goodwill',
       'GrossProfit', 'IncreaseDecreaseInInventories', 'InterestExpense',
       'InventoryNet', 'Liabilities', 'LiabilitiesAndStockholdersEquity',
       'LiabilitiesCurrent', 'LongTermDebt', 'LongTermDebtCurrent',
       'LongTermDebtMaturitiesRepaymentsOfPrincipalInNextTwelveMonths',
       'LongTermDebtNoncurrent', 'NetCashProvidedByUsedInFinancingActivities',
       'NetCashProvidedByUsedInInvestingActivities',
       'NetCashProvidedByUsedInOperatingActivities', 'NetIncomeLoss',
       'OperatingExpenses', 'OperatingIncomeLoss', 'ProfitLoss',
       'RepaymentsOfLongTermDebt', 'Revenues', 'SalesRevenueGoodsNet',
       'SalesRevenueNet', 'StockholdersEquity', 'TotalAsset',
       'WorkingCapital']


f1_wide_sec.update(f1_wide_sec.groupby(['company_name', 'period_end_date'])[cols].ffill())

f1_wide_sec.head()

# If assume companies that have some missing fields have balanced financials, can fill in the missing fields

# fill in StockholdersEquity 

f1_wide_sec['StockholdersEquity'].fillna(f1_wide_sec['LiabilitiesAndStockholdersEquity'] - f1_wide_sec['Liabilities'], inplace=True)

f1_wide_sec['StockholdersEquity'].isna().value_counts()

## fill in LiabilitiesAndStockholdersEquity

f1_wide_sec['LiabilitiesAndStockholdersEquity'].fillna(f1_wide_sec['StockholdersEquity'] + f1_wide_sec['Liabilities'], inplace=True)

f1_wide_sec['LiabilitiesAndStockholdersEquity'].isna().value_counts()

## fill in Liabilities

f1_wide_sec['Liabilities'].fillna(f1_wide_sec['LiabilitiesAndStockholdersEquity'] - f1_wide_sec['StockholdersEquity'], inplace=True)

f1_wide_sec['Liabilities'].isna().value_counts()

# couple companies reported Assets in Total Assets

f1_wide_sec['Assets'].fillna(f1_wide_sec['TotalAsset'], inplace=True)

f1_wide_sec['Assets'].isna().value_counts()

# majority of companies use the AssetsCurrent field.  Few used CurrentAsset

f1_wide_sec['AssetsCurrent'].fillna(f1_wide_sec['CurrentAsset'], inplace=True)

f1_wide_sec.drop(['CurrentAsset', 'TotalAsset'], axis=1, inplace=True)

# filter to where Assets >= 100MM
# this is the threshold for the bankruptcy list from UCLA

f2_wide_sec = f1_wide_sec[f1_wide_sec['Assets'] >= 1e8]
f2_wide_sec.shape

f2_wide_sec.info()

# df already sorted by company_name, period_end_date, and date_filed
# and has been filled forward within company_name, period_end_date group

# keep only one row for company_name, period_end_date

# no company should have more than 5 rows, but this may have occured if 
# data submission labeled 10-Q rather than 10-K 


f3_wide_sec  = f2_wide_sec.groupby(['company_name', 'period_end_date']).tail(1)
f3_wide_sec.shape

f3_wide_sec['company_name'].value_counts()

f3_wide_sec[f3_wide_sec['company_name'] == 'PENNYMAC FINANCIAL SERVICES, INC.' ].head()

## after examining a few companies, their quarterly files are with the annual form 10-K
## just filtering on number_of_quarters = 4 will remove good data
## even the largest companies with 1 row per year, have some number_of_quarters=0 for their annual report

f3_wide_sec.fiscal_year_end.unique()

## the fiscal_year_end field could be used to apply logic 
## to keep rows where month and day of period_end_date
## equals the month and day of fiscal_year_end

f3_wide_sec['FY_end'] = f3_wide_sec['fiscal_year_end'].astype(str)
f3_wide_sec['FY_end_day'] = f3_wide_sec['FY_end'].str[-2:].astype(int)
f3_wide_sec['FY_end_month'] = f3_wide_sec['FY_end'].str[:-2].astype(int)
f3_wide_sec.head()

f3_wide_sec['keep_row'] = np.where((f3_wide_sec['period_end_date'].dt.month ==f3_wide_sec['FY_end_month'] )
                                   & (f3_wide_sec['period_end_date'].dt.day ==f3_wide_sec['FY_end_day'] ), 1, 0)

f3_wide_sec.head()

f3_wide_sec.shape


f4_wide_sec = f3_wide_sec[f3_wide_sec['keep_row']==1]
f4_wide_sec.shape

################################

# add the industry descriptions for each SIC code


query_ind = "SELECT * FROM sic_codes;"

sic_codes = pd_sql.read_sql(query_ind, connection)
sic_codes.columns

query = """ 
SELECT a.*, b.Industry
FROM f4_wide_sec as a 
    LEFT JOIN sic_codes as b
ON a.sic = b.[SIC Code]
"""

ind_wide_sec = pysqldf(query)
ind_wide_sec.head()

## this join made period_end_date and date_filed into datetimes

ind_wide_sec.shape

# add overall industry division discription

ind_wide_sec['div_code'] = ind_wide_sec['sic'].str[:2]

query_div = "SELECT * FROM sic_divisions;"

sic_div = pd_sql.read_sql(query_div, connection)
sic_div.columns

query = """ 
SELECT a.*, b.Division
FROM ind_wide_sec as a 
    LEFT JOIN sic_div as b
ON a.div_code = b.div_code
"""

div_wide_sec = pysqldf(query)
div_wide_sec.head()


div_wide_sec.shape

div_wide_sec['Division'].value_counts()

div_wide_sec['Division'].count()

div_wide_sec.info()

# save the wide table with industry descriptions in postgres db

from sqlalchemy import create_engine
engine = create_engine('postgresql://amybutler:localhost@localhost:5432/bankruptcy')

div_wide_sec.to_sql('sec_wide_table', engine, index=False)

#check table saved and can be queried
query = 'SELECT * FROM sec_wide_table LIMIT 5;'
df = pd.read_sql(query, engine)

df.head()
