==> code/06_test_SEC_BRD_merge.ipynb <==
""" 
examine why brd list contents didn't all match up to
SEC > 100M  annual reports
"""

import psycopg2 as pg
import pandas.io.sql as pd_sql
import pandas as pd
import numpy as np
import datetime


# Postgres info to connect

connection_args = {
    'host': 'localhost',  
    'dbname': 'bankruptcy',  
    'port': 5432          
}

connection = pg.connect(**connection_args)
connection.autocommit = True   
                              

# few fields for first initial model
# all 2019 data held out for test set

query = """
SELECT *
FROM sec_wide_table
WHERE  period_end_date < '2019-01-01' 
;
"""

sec_data = pd_sql.read_sql(query, connection)
sec_data.head()

company_unique = pd.DataFrame(sec_data['company_name'].unique(), columns={'Name'})

company_unique = set(sec_data['company_name'].unique())
company_unique

#brd_list
brd_labels = pd.read_excel('debtor_list_ein_lookup.xlsx', sheet_name='LIST')
brd_labels.head()

#capitalize 
brd_labels['company_name_brd'] = brd_labels['name_in_sec_data'].str.upper()
brd_labels.head()

brd_names = set(brd_labels['company_name_brd'])
brd_names

# compare the two sets 

brd_names.difference(company_unique)

query = """
SELECT distinct company_name, period_end_date
FROM sec_all_raw
WHERE  period_end_date < '2019-01-01' 
    and company_name LIKE '%BLACK ELK ENERGY%'
;
"""

one_company = pd_sql.read_sql(query, connection)
one_company

#black_elk filed for bankruptcy on 8/11/2015  
# has 10-Q 2014-09-30

query = """
SELECT distinct company_name, form, number_of_quarters, period_end_date
FROM sec_raw_all
WHERE  period_end_date < '2019-01-01' 
    and company_name LIKE '%CAL DIVE INTERNATIONAL, INC.%'
;
"""

one_company = pd_sql.read_sql(query, connection)
one_company

# filed 3/3/2015 
# 2014 annual report would have been due in April 2015
# company filed for bankruptcy before the annual report due
# but did they file quarterly reports? 
# yes, turned in a 2014-09-30 10-Q


query = """
SELECT distinct company_name, form, number_of_quarters, period_end_date
FROM sec_raw_all
WHERE  period_end_date < '2019-01-01' 
    and company_name LIKE '%STANDARD REGISTER%'
;
"""

one_company = pd_sql.read_sql(query, connection)
one_company


## have quarterly report fall 2014, bankruptcy 03/12/2015




query = """
SELECT distinct company_name, form, number_of_quarters, period_end_date
FROM sec_raw_all
WHERE  period_end_date < '2019-01-01' 
    and company_name LIKE '%SUNEDISON, INC%'
;
"""

one_company = pd_sql.read_sql(query, connection)
one_company

# bankruptcy 4/21/2016
# have 10-Q 	2015-09-30


query = """
SELECT distinct company_name, form, number_of_quarters, period_end_date
FROM sec_raw_all
WHERE  period_end_date < '2019-01-01' 
    and company_name LIKE '%VENOCO%'
;
"""

one_company = pd_sql.read_sql(query, connection)
one_company
#bankruptcy 3/18/2016  - some filings are VENOCO other filings are DENVER PARENT CORP
#BRD list has Venoco, Inc. (Denver Parent Corp.)
    


query = """
SELECT distinct company_name, form, number_of_quarters, period_end_date
FROM sec_raw_all
WHERE  period_end_date < '2019-01-01' 
    and company_name LIKE '%DENVER PARENT%'
;
"""

one_company = pd_sql.read_sql(query, connection)
one_company
 #bankruptcy 3/18/2016  - some filings are VENOCO
