==> code/11_SEC_2020Q1_10K.ipynb <==
"""
Annual 10-K filing for Q4 2019 are not in the Google Cloud Platform database

Text files of data available by quarter from SEC website

Many companies sent 2019 reports in Q1 2020 

prepare Q1 2020 data and extract the 2019 period_end_date reports
"""

import pandas as pd
import numpy as np
import datetime as dt

import pandas.io.sql as pd_sql

from pandasql import sqldf
# PandaSQL needs to be able to reference the global variables already defined
pysqldf = lambda q: sqldf(q, globals())

import psycopg2 as pg

# Postgres info to connect

connection_args = {
    'host': 'localhost',  
    'dbname': 'bankruptcy',  
    'port': 5432          
}

connection = pg.connect(**connection_args)
connection.autocommit = True   

#The NUM data set contains numeric data, one row per data point in the financial statements.
NUM_q = pd.read_csv('../data/2020q1/num.txt', delimiter='\t' , 
                     parse_dates=['ddate'] )

NUM_q.shape

NUM_q.head()

# detailed explaination of each field provided by SEC
# update field names to match the GCP labels

NUM_q.columns = ['adsh', 'measure_tag', 'version', 'coreg', 
                  'period_end_date', 'number_of_quarters','units', 'value','footnote']

NUM_q.head()

NUM_q.info()

#The SUB (submissions) data set contains summary information about an entire EDGAR submission. 

SUB_q = pd.read_csv('../data/2020q1/sub.txt', delimiter='\t', 
                     parse_dates=['period'], 
                     dtype={'sic':  'str', 'ein':'str', 'cik':'str', 'fy':'str', 'fp':'str'})

SUB_q.shape

SUB_q.head()

# check forms included in this file
# only keep the 10-K and 10-K/A

SUB_q.form.unique()

# check all US located companies
# will drop non-US companies

SUB_q.countryba.unique()

SUB_q.columns

SUB_q.info()

# Will merge NUM and SUB files and then filter like how GCP queries were filtered
# per details from SEC of how to merge these datasets on adsh

all_q = pd.merge(NUM_q, SUB_q, how='left', on=['adsh'])

all_q.shape

#all_q.head()

#all_q.tail()

#all_q.columns

import pickle

with open('Q1_2020.pickle', 'wb') as to_write:
    pickle.dump(all_q, to_write)
