==> code/14_prepare_SEC_2019.ipynb <==
"""

Extra text files had to be downloaded from SEC website because
annual 10-K filing for Q4 2019 are not in the Google Cloud Platform database

these files were merged and prepared to be like extracts from Google Cloud SEC database

stack all pickled files together and further refine data to one 2019 report per company

saved prepared 2019 financial report data to local postgres db

TEST HOLD OUT DATA
"""

import pandas as pd
import numpy as np
import datetime as dt

import pandas.io.sql as pd_sql

from pandasql import sqldf
# PandaSQL needs to be able to reference the global variables already defined
pysqldf = lambda q: sqldf(q, globals())

import pickle

import psycopg2 as pg

# Postgres info to connect

connection_args = {
    'host': 'localhost',  
    'dbname': 'bankruptcy',  
    'port': 5432          
}

connection = pg.connect(**connection_args)
connection.autocommit = True   

with open('q4_2019.pickle', 'rb') as read_file:
    q4_2019 = pickle.load(read_file)
    
q4_2019.shape

with open('q1_2020.pickle', 'rb') as read_file:
    q1_2020 = pickle.load(read_file)
    
q1_2020.shape

combined = pd.concat([q4_2019,q1_2020], axis=0 )
combined.shape

## stack with Q2 2020

with open('q2_2020.pickle', 'rb') as read_file:
    q2_2020 = pickle.load(read_file)
    
q2_2020.shape


combined2 = pd.concat([combined,q2_2020], axis=0 )
combined2.shape

## stack with Q3 2020

with open('Q3_2020.pickle', 'rb') as read_file:
    Q3_2020 = pickle.load(read_file)
    
Q3_2020.shape

combined3 = pd.concat([combined2,Q3_2020], axis=0 )
combined3.shape

combined3.columns

########################

# apply filters to keep only 10-K reports from US companies
f1_all_q1 = combined3[combined3['countryba'] == 'US']
f1_all_q1.shape

f2_all_q1 = f1_all_q1[f1_all_q1['form'].isin(['10-K', '10-K/A'])]
f2_all_q1.shape

# fp is Fiscal Period Focus.  The 10-K is to have a fiscal period focus of FY
f2_all_q1.fp.unique()

# keep only the columns that match GCP fields
# rename columns 

f3_all_q1 = f2_all_q1[['name', 'period', 'number_of_quarters', 'adsh', 'cik', 'ein', 
           'sic', 'fy', 'form', 'filed', 'measure_tag', 'value', 'units' ]]
f3_all_q1.shape

f4_all_q1 = f3_all_q1[f3_all_q1['period'] > dt.datetime(2018,12,31)]

f4_all_q1.head()


f4_all_q1.shape

max(f4_all_q1.period)

f4_all_q1.period.value_counts()

# rename columns so will match GCP labeling

f4_all_q1.columns = ['company_name', 'period_end_date', 'number_of_quarters',
                    'submission_number', 'central_index_key', 'ein',
                    'sic', 'fiscal_year', 'form', 'date_filed',
                    'measure_tag', 'value', 'units']

f4_all_q1.head()

f4_all_q1.company_name.nunique()

#f4_all_q1[f4_all_q1.company_name == 'HORNBECK OFFSHORE SERVICES INC /LA']

# filter to only measurements queried and downloaded from GCP

query = """ 
SELECT *
FROM f4_all_q1 as a 
where measure_tag in ('Assets',
       'AssetsCurrent', 'CashAndCashEquivalentsAtCarryingValue',
       'CashAndCashEquivalentsPeriodIncreaseDecrease', 'CommonStockValue',
       'CostsAndExpenses', 'CurrentAsset', 'DeferredIncomeTaxExpenseBenefit',
       'DeferredIncomeTaxLiabilities', 'DeferredIncomeTaxLiabilitiesNet',
       'DeferredIncomeTaxesAndTaxCredits', 'Depreciation',
       'EarningsPerShareBasic', 'EarningsPerShareDiluted', 'Goodwill',
       'GrossProfit', 'IncreaseDecreaseInInventories', 'InterestExpense',
       'InventoryNet', 'Liabilities', 'LiabilitiesAndStockholdersEquity',
       'LiabilitiesCurrent', 'LongTermDebt', 'LongTermDebtCurrent',
       'LongTermDebtMaturitiesRepaymentsOfPrincipalInNextTwelveMonths',
       'LongTermDebtNoncurrent', 'NetCashProvidedByUsedInFinancingActivities',
       'NetCashProvidedByUsedInInvestingActivities',
       'NetCashProvidedByUsedInOperatingActivities', 'NetIncomeLoss',
       'OperatingExpenses', 'OperatingIncomeLoss', 'ProfitLoss',
       'RepaymentsOfLongTermDebt', 'Revenues', 'SalesRevenueGoodsNet',
       'SalesRevenueNet', 'StockholdersEquity', 'TotalAsset',
       'WorkingCapital')
"""

f5_all_q1 = pysqldf(query)
f5_all_q1.head()



f5_all_q1.shape

# check all units in USD
f5_all_q1.units.unique()

# keep only USD,  as found in EDA, some non-US companies file the US form and no requirement they don't

f6_all_q1 = f5_all_q1[f5_all_q1['units'] == 'USD']
f6_all_q1.shape

# df is tall on each measurement 
# pivot to wide


wide_sec = f6_all_q1.pivot_table(index=['company_name','period_end_date', 'number_of_quarters',
                                              'submission_number','central_index_key',
                                             'ein', 'sic', 'fiscal_year', 'form',
                                             'date_filed'],
                        columns='measure_tag', values='value').reset_index()
wide_sec.head()


wide_sec.shape

wide_sec.company_name.nunique()

##################

# combine these annual reports with GCP 2019 file download that contains 1/1/2019 - 9/30/2019 priod end dates


query = """
SELECT *
FROM sec_all_usd
WHERE  period_end_date > '2018-12-31' 
;
"""

gcp_data = pd_sql.read_sql(query, connection)
gcp_data.head()

gcp_data.shape

gcp_data.company_name.nunique()

gcp_data.columns

# pivot wide

wide_gcp = gcp_data.pivot_table(index=['company_name','period_end_date', 'number_of_quarters',
                                              'submission_number','central_index_key',
                                             'ein', 'sic', 'fiscal_year_end', 'fiscal_year', 'form',
                                             'date_filed'],
                        columns='measure_tag', values='value').reset_index()
wide_gcp.head()

wide_gcp.shape

# stack the 2 df with 2019 financial reports

sec_2019 = pd.concat([wide_gcp, wide_sec], axis=0)
sec_2019.shape

sec_2019[['company_name', 'period_end_date']].value_counts()

# like with handling 2014-2018 data, will keep where quarters zero or 4
# then forward fill and back fill
# will filter to have only one row for company and period_end_date



sec_df2 = sec_2019[sec_2019['number_of_quarters'].isin([0,4])]
sec_df2.shape

# fill forward within groupby, then back fill within groupby

# leave the NaN's to apply additional logics

cols = ['Assets',
       'AssetsCurrent', 'CashAndCashEquivalentsAtCarryingValue',
       'CashAndCashEquivalentsPeriodIncreaseDecrease', 'CommonStockValue',
       'CostsAndExpenses', 'CurrentAsset', 'DeferredIncomeTaxExpenseBenefit',
       'DeferredIncomeTaxLiabilities', 'DeferredIncomeTaxLiabilitiesNet',
       'DeferredIncomeTaxesAndTaxCredits', 'Depreciation',
       'EarningsPerShareBasic', 'EarningsPerShareDiluted', 'Goodwill',
       'GrossProfit', 'IncreaseDecreaseInInventories', 'InterestExpense',
       'InventoryNet', 'Liabilities', 'LiabilitiesAndStockholdersEquity',
       'LiabilitiesCurrent', 'LongTermDebt', 'LongTermDebtCurrent',
       'LongTermDebtMaturitiesRepaymentsOfPrincipalInNextTwelveMonths',
       'LongTermDebtNoncurrent', 'NetCashProvidedByUsedInFinancingActivities',
       'NetCashProvidedByUsedInInvestingActivities',
       'NetCashProvidedByUsedInOperatingActivities', 'NetIncomeLoss',
       'OperatingExpenses', 'OperatingIncomeLoss', 'ProfitLoss',
       'RepaymentsOfLongTermDebt', 'Revenues', 'SalesRevenueGoodsNet',
       'SalesRevenueNet', 'StockholdersEquity', 
       'WorkingCapital']


sec_df2.update(sec_df2.groupby(['company_name', 'period_end_date'])[cols].ffill())


sec_df2.update(sec_df2.groupby(['company_name', 'period_end_date'])[cols].bfill())

# like done for 2014 - 2018
# If assume companies that have some missing fields have balanced financials, can fill in the missing fields

f1_wide_sec = sec_df2.copy()

# fill in StockholdersEquity 

f1_wide_sec['StockholdersEquity'].fillna(f1_wide_sec['LiabilitiesAndStockholdersEquity'] - f1_wide_sec['Liabilities'], inplace=True)

f1_wide_sec['StockholdersEquity'].isna().value_counts()

## fill in LiabilitiesAndStockholdersEquity

f1_wide_sec['LiabilitiesAndStockholdersEquity'].fillna(f1_wide_sec['StockholdersEquity'] + f1_wide_sec['Liabilities'], inplace=True)

f1_wide_sec['LiabilitiesAndStockholdersEquity'].isna().value_counts()

## fill in Liabilities

f1_wide_sec['Liabilities'].fillna(f1_wide_sec['LiabilitiesAndStockholdersEquity'] - f1_wide_sec['StockholdersEquity'], inplace=True)

f1_wide_sec['Liabilities'].isna().value_counts()

# majority of companies use the AssetsCurrent field.  Few used CurrentAsset

f1_wide_sec['AssetsCurrent'].fillna(f1_wide_sec['CurrentAsset'], inplace=True)

f1_wide_sec['Assets'].isna().value_counts()

## Different companies put Revenues in different fields

f1_wide_sec['Revenue_any'] = f1_wide_sec['Revenues']

f1_wide_sec['Revenue_any'].fillna(f1_wide_sec['SalesRevenueNet'], inplace=True)
f1_wide_sec['Revenue_any'].fillna(f1_wide_sec['SalesRevenueGoodsNet'], inplace=True)

f1_wide_sec.columns

f1_wide_sec.head()

# sort df
f1_wide_sec.sort_values(by=['company_name', 'period_end_date'], inplace=True)
f1_wide_sec.head()

#date_filed

# filter to where Assets >= 100MM
# this is the threshold for the bankruptcy list from UCLA

f2_wide_sec = f1_wide_sec[f1_wide_sec['Assets'] >= 1e8]
f2_wide_sec.shape

f2_wide_sec.info()

# period_end_date back to a datetime

f2_wide_sec['period_end_date']= pd.to_datetime(f2_wide_sec['period_end_date'])

f3_wide_sec  = f2_wide_sec.groupby(['company_name', 'period_end_date']).tail(1)
f3_wide_sec.shape

f3_wide_sec['company_name'].value_counts()

f4_wide_sec = f3_wide_sec[f3_wide_sec['period_end_date'] < dt.datetime(2020,1,1)]

f4_wide_sec.shape

f4_wide_sec.company_name.nunique()

#########
# add industry and industry division

query_ind = "SELECT * FROM sic_codes;"

sic_codes = pd_sql.read_sql(query_ind, connection)
sic_codes.columns

ind_wide_sec = pd.merge(f4_wide_sec, sic_codes, how='left', left_on='sic', right_on='SIC Code' )

ind_wide_sec.head()

ind_wide_sec.shape

# add overall industry division discription

ind_wide_sec['div_code'] = ind_wide_sec['sic'].str[:2]

query_div = "SELECT * FROM sic_divisions;"

sic_div = pd_sql.read_sql(query_div, connection)
sic_div.columns


div_wide_sec = pd.merge(ind_wide_sec, sic_div, how='left', on='div_code')

div_wide_sec.shape

div_wide_sec['Division'].value_counts()

div_wide_sec['Division'].count()

## then to db table? 

from sqlalchemy import create_engine
engine = create_engine('postgresql://amybutler:localhost@localhost:5432/bankruptcy')

div_wide_sec.to_sql('sec_prep_2019', engine, index=False, if_exists='replace')

#check table saved and can be queried
query = 'SELECT * FROM sec_prep_2019 LIMIT 5;'
df = pd.read_sql(query, engine)

df.head()


with open('sec_test_data.pickle', 'wb') as to_write:
    pickle.dump(wide_sec, to_write)
