==> code/18_merge_SEC_BRD_test_data.ipynb <==
"""
query SEC data cleaned and prepared for 2019 and saved in postgres db

Merge with prepared bankruptcy list

TEST HOLD OUT DATA
"""

import psycopg2 as pg
import pandas as pd
import pandas.io.sql as pd_sql
import numpy as np
import datetime

from pandasql import sqldf
# PandaSQL needs to be able to reference the global variables already defined (namely, `df`)
pysqldf = lambda q: sqldf(q, globals())

# Postgres info to connect

connection_args = {
    'host': 'localhost',  
    'dbname': 'bankruptcy',  
    'port': 5432          
}

connection = pg.connect(**connection_args)
connection.autocommit = True   
                              

# few fields for first initial model
# all 2019 data held out for test set

query = """
SELECT *
FROM sec_prep_2019
;
"""

sec_data = pd_sql.read_sql(query, connection)
sec_data.head()


sec_data.columns

sec_data.info()

sec_data.shape

# add the labels for the target
# sheet LIST contains 2015-2019 bankruptcies where company names where adjusted 
# to be exactly like in SEC filings

brd_list = pd.read_excel('debtor_list_ein_lookup.xlsx', sheet_name='FULL_LIST')
#brd_labels

brd_list['FULL_NAME'] = brd_list['name_in_sec_data'].str.upper()
brd_list['bankruptcy_date'] = brd_list['date_filed']
brd_list.head()

brd_list.shape

brd_list['FULL_NAME'].nunique()

# Chaparral Energy filed in 2016 and 2020
# Ultra Petroleum Corp filed in 2016 and 2020
# American Apparel
# Geophysical
# Halcon Resources
# Paragon Offshore

brd_labels = brd_list[['FULL_NAME', 'bankruptcy_date']]

add_brd = pd.merge(sec_data, brd_labels, how='left', left_on='company_name', right_on='FULL_NAME')


add_brd.head()

add_brd.shape

add_brd[add_brd['bankruptcy_date'].notnull()].head()

#add_brd.info()

#put back to a datetime variable

add_brd['period_end_date'] = pd.to_datetime(add_brd['period_end_date'])


## if bankruptcy date is within 1 year of period_end_date , then target = 1

## add column that calculates difference in 2 dates

add_brd['time_delta'] =  add_brd['bankruptcy_date'] - add_brd['period_end_date'] 
add_brd['days'] = add_brd["time_delta"].dt.days
add_brd.head()

#add_brd['target'] = np.where((add_brd['days'] >= 0) & (add_brd['days'] <=365) , 1, 0)
add_brd['target'] = np.where(add_brd['bankruptcy_date'] >= datetime.datetime(2020,1,1), 1, 0)

add_brd[add_brd['bankruptcy_date'].notnull()].head()

add_brd[add_brd['target'] == 1].head()

add_brd.target.value_counts()

# Searched for filings in SEC EDGAR
# Three companies did not send a 2019 report to SEC

# Neiman Marcus
# Centric Brands
# AAC Holdings, INC

add_brd.company_name.nunique()

## examine further

add_brd.to_excel('review_test_df.xlsx', index=False)

# pickle the model_data df
import pickle

with open('test_data.pickle', 'wb') as to_write:
    pickle.dump(add_brd, to_write)
